+++ src/Core/NSFileManagerOOExtensions.m
@@ -235,7 +235,7 @@ MA 02110-1301, USA.
 
 - (BOOL) oo_oxzFileExistsAtPath:(NSString *)path
 {
-	unsigned i, cl;
+	NSUInteger i, cl;
 	NSArray *components = [path pathComponents];
 	cl = [components count];
 	for (i = 0 ; i < cl ; i++)
